Refactored from sga.core + new interface for lazy reading/writing
The old method exposed a lot of extra properties on classes which didn't need to expose that; since their job was to read/write sections of the stream.

The new method uses composition instead of inheritance; but allows for inheritance chains to continue to work by using the BinaryProxy protocol.

This primarily allows the 'serializers' to continue to act as BinaryIO objects without exposing all the BinaryIO info