fix(lazyio): CString padding
When we handle padding is extremely important due to encoding using different sized bytes
By using a padding string, and stripping after decoding, we ensure that decodes dont error due to malformed byte buffers.
Now we should only see bad buffers IF the cstring buffer size does not align to the encoding size, or a variable-length encoding is used.